This PR:
- removes `_tmp_dissect_pos`
- switch unecessary deepcopy to copy
- store a struct object within the field to avoid recompiling it
- previously, `do_dissect` would convert any -> bytes.  This was implemented as safety workaround at the very beggining of Python 3 support. Some protocols / tests (DPE_RPC, ISOTP, others..) abused this behavior to pass wrongly built strings. Remove this behavior

Perf: using `test/benchmark/dissection_and_build.py` - `N=10000`

| Type | Before - 2.7|  Before - 3.6 | After - 2.7 | After - 3.6 |
| --- | --- | --- | --- | --- |
| Build | 6.15s | 4.48s | 6.12s (~) | 4.46s (~) |
| Dissect | 1.94s | 1.59s | 1.72s (+11%) | 1.40s (+11%) |
| Build & dissect | 8.62s | 6.26s | 8.19s (+4%) | 5.97s (+4%) |

Note: build times are slightly less consistent than dissection times.