@guedou 2019-12-15
Could you share the performance improvements using the script from #642 ?