@gpotter2 2019-12-22
@guedou I've updated my first post